third_party/blink/web_tests/printing/page-rule-selection-expected.txt

This tests page style selection.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Check font update and line height override.
PASS internals.pageProperty('line-height', 0) is "180"
PASS internals.pageProperty('font-family', 0) is "Arial"
PASS internals.pageProperty('font-size', 0) is "30"
Check @media rules.
PASS internals.pageProperty('size', 0) is "100 200"
PASS internals.pageProperty('size', 0) is "150 250"
PASS internals.pageProperty('size', 0) is "150 250"
PASS successfullyParsed is true

TEST COMPLETE
